[CLOUD-2730] Incorporate Microprofile Health checks in the liveness / readiness probes.
Both DMR and jolokia probes are added.

For the DMR probe, use a separate probe class for the MP Health check as combining the calls to the health subsystem as steps in the same composite as the existing ops for other tests cannot be made to work reliably across different EAP versions. The composite will fail in different ways if the health extension is not present, but the overall probe must work reliably whether or not the extension is there (and it won't be for EAP 6.x and EAP 7.1.x or anything based on EAP that doesn't include it.)

Also,  DMR probes should handle failures in 1 step in the composite without assuming all steps are failed.

Signed-off-by: Brian Stansberry <brian.stansberry@redhat.com>