Rename/move maven modules to provide versions alternatives
The module name "jboss.container.maven.35.bash" is now provided by two
modules, at jboss/container/maven/35/{default,35}. The default one is
the older SCL-based one that is in use by many images. The '35/' path
is a simpler one for when the maven package is available without SCL,
simply as "maven".

For the OpenJDK image at least, we can force the use of the newer module
by depending upon it in an overrides file and setting the version
requirement accordingly. By default, the "default" module *appears* to
be chosen by Cekit's module resolver, so the older images get the right
one automatically. If this was found to not be true in some
circumstances the old module could be forced using the same technique
as the new one.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>