enable Maven 3.6 AppStream module prior to install
The AppStream module part of this commit is necessary to get access
to Maven 3.6 (instead of 3.5). The "dnf module enable" part needs
to happen before the package installation. Cekit's module ordering
places the "execute" last; so we either need to have two cekit
modules: the first to enable the AppStream module, and the second
to install the package; or (as in this commit) perform the package
installation in the script, too.

This hard-codes "dnf". However, "microdnf" does not support enabling
AppStream modules, which rules out using the minimal-flavour images
without further work.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>