revamp rm-openjdk module as singleton-jdk
Infer the correct Java packages that should be in the container
from the environment, specifically JAVA_VERSION and JAVA_VENDOR.

Remove similar code from the various Maven modules.

Images which use the Maven modules should depend on "singleton-jdk"
explicitly and ensure that it is sequenced after the Maven modules
or any other module that might result in the installation of extra
Java packages (e.g. jolokia…)

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>