Enable maven AppStream module via conf file
By providing /etc/dnf/modules.d/maven.module before attempting to
install Maven, we can get access to Maven 3.6 from the AppStream
repository without having to enable the module with full-fat DNF.
This means we can install Maven 3.6 with microdnf.

The maven.module file has to be in-place before attempting to install
Maven. Since CeKit processes "packages:" before "execute:" within the
context of a module, we need to put the new "execute:" stuff in a new
module that is depended upon by those that use "packages:" to install
maven.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>